// types.ts
export type ResourceType = 'book' | 'flashcard' | 'notes' | 'qa' | 'activity' | 'video' | 'audio' | 'worksheet' | 'questionPaper' | 'quiz' | 'slide';